Move models to a package
